rename tests to _tests to run long tests first
