+++ sktime/classification/tests/test_all_classifiers.py
@@ -12,7 +12,7 @@ import numpy as np
 import pytest
 
 from sktime.registry import all_estimators
-from sktime.tests._config import EXCLUDE_ESTIMATORS, NON_STATE_CHANGING_METHODS
+from sktime._tests._config import EXCLUDE_ESTIMATORS, NON_STATE_CHANGING_METHODS
 from sktime.utils._testing.estimator_checks import _has_capability, _make_args
 
 CLASSIFIERS = all_estimators(
