+++ sktime/regression/tests/test_all_regressors.py
@@ -13,7 +13,7 @@ import pandas as pd
 import pytest
 
 from sktime.registry import all_estimators
-from sktime.tests._config import EXCLUDE_ESTIMATORS, NON_STATE_CHANGING_METHODS
+from sktime._tests._config import EXCLUDE_ESTIMATORS, NON_STATE_CHANGING_METHODS
 from sktime.utils._testing.estimator_checks import _has_capability, _make_args
 
 REGRESSORS = all_estimators(
