+++ sktime/utils/_testing/estimator_checks.py
@@ -22,7 +22,7 @@ from sktime.datatypes._panel._check import is_nested_dataframe
 from sktime.dists_kernels import BasePairwiseTransformer, BasePairwiseTransformerPanel
 from sktime.forecasting.base import BaseForecaster
 from sktime.regression.base import BaseRegressor
-from sktime.tests._config import VALID_ESTIMATOR_TYPES
+from sktime._tests._config import VALID_ESTIMATOR_TYPES
 from sktime.transformations.base import (
     BaseTransformer,
     _PanelToPanelTransformer,
